refactor tests: organize apitest files (fix #4)
